move to save

+++ src/io/iologindata.cpp
@@ -1085,34 +1085,11 @@ bool IOLoginData::savePlayer(Player* player)
     }
   }
 
-  //save reward items
-  query.str(std::string());
-  query << "DELETE FROM `player_rewards` WHERE `player_id` = " << player->getGUID();
-
-  if (!db.executeQuery(query.str())) {
-    return false;
+  if(!IOLoginDataSave::saveRewardItems(player)){
+	  return false;
   }
 
-  std::vector<uint64_t> rewardList;
-  player->getRewardList(rewardList);
-
-  if (!rewardList.empty()) {
-    DBInsert rewardQuery("INSERT INTO `player_rewards` (`player_id`, `pid`, `sid`, `itemtype`, `count`, `attributes`) VALUES ");
-    itemList.clear();
-
-    for (const auto& rewardId : rewardList) {
-      Reward* reward = player->getReward(rewardId, false);
-      if (!reward->empty() && ((time(nullptr) - rewardId / 1000) <= 60 * 60 * 24 * 7)) {
-        itemList.emplace_back(0, reward);
-      }
-    }
-
-    if (!saveItems(player, itemList, rewardQuery, propWriteStream)) {
-      return false;
-    }
-  }
-
-  //save inbox items
+	//save inbox items
   query.str(std::string());
   query << "DELETE FROM `player_inboxitems` WHERE `player_id` = " << player->getGUID();
   if (!db.executeQuery(query.str())) {
